hf-classifier/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.30.0
numpy>=1.24.0
Pillow>=10.0.0
torch>=2.0.0
torchvision>=0.15.0
opencv-python>=4.8.0

